Get rid of possible dangling popup view (always remove from super view)
The first condition: if(!self.superview){...} inhibits the view from being added to the window if it had a superview, so it is assuming that the superview is always the view showing that is showing now, which might not be the case.
To make sure we never have a dangling PopupView: we remove the PopupView from it's superview then continue our process.